+++ ios/PiwikProSdk.m
@@ -146,7 +146,24 @@ RCT_REMAP_METHOD(trackApplicationInstall,
     }
     
     @try {
-        [[PiwikTracker sharedInstance] sendApplicationDownload];
+      [[PiwikTracker sharedInstance] applicationInstall];
+        resolve(nil);
+    } @catch (NSException *exception) {
+        reject(exception.name, exception.reason, nil);
+    }
+}
+
+RCT_REMAP_METHOD(trackApplicationUpdate,
+                 trackApplicationUpdateWithResolver:(RCTPromiseResolveBlock)resolve
+                 withRejecter:(RCTPromiseRejectBlock)reject)
+{
+    if ([PiwikTracker sharedInstance] == nil) {
+        reject(@"not_initialized", @"Piwik Pro SDK has not been initialized", nil);
+        return;
+    }
+    
+    @try {
+      [[PiwikTracker sharedInstance] applicationUpdate];
         resolve(nil);
     } @catch (NSException *exception) {
         reject(exception.name, exception.reason, nil);
@@ -252,7 +269,7 @@ RCT_REMAP_METHOD(trackGoal,
     @try {
         [self applyOptionalParameters:options];
         
-        [[PiwikTracker sharedInstance] sendGoalWithID:goal revenue:options[@"revenue"]];
+        [[PiwikTracker sharedInstance] sendGoalWithID:goal revenue:options[@"revenue"] currencyCode:options[@"currencyCode"]];
         resolve(nil);
     } @catch (NSException *exception) {
         reject(exception.name, exception.reason, nil);
@@ -303,8 +320,8 @@ RCT_REMAP_METHOD(trackEcommerceProductDetailView,
     @try {
         [self applyOptionalParameters:options];
         EcommerceProducts *ecommerceProducts = [self buildEcommerceProducts:products];
-        [[PiwikTracker sharedInstance] ecommerceProductDetailView: ecommerceProducts];
-
+        [[PiwikTracker sharedInstance] ecommerceProductDetailView:ecommerceProducts
+                                                    currencyCode: options[@"currencyCode"]];
         resolve(nil);
     } @catch (NSException *exception) {
         reject(exception.name, exception.reason, nil);
@@ -326,7 +343,8 @@ RCT_REMAP_METHOD(trackEcommerceCartUpdate,
     @try {
         [self applyOptionalParameters:options];
         EcommerceProducts *ecommerceProducts = [self buildEcommerceProducts:products];
-        [[PiwikTracker sharedInstance] ecommerceCartUpdate: ecommerceProducts grandTotal: grandTotal];
+        [[PiwikTracker sharedInstance] ecommerceCartUpdate: ecommerceProducts grandTotal: grandTotal
+                                              currencyCode: options[@"currencyCode"]];
 
         resolve(nil);
     } @catch (NSException *exception) {
@@ -347,8 +365,9 @@ RCT_REMAP_METHOD(trackEcommerceAddToCart,
 
     @try {
         [self applyOptionalParameters:options];
-        EcommerceProducts *ecommerceProducts = [self buildEcommerceProducts:products];
-        [[PiwikTracker sharedInstance] ecommerceAddToCart: ecommerceProducts];
+        EcommerceProducts *ecommerceProducts = [self buildEcommerceProducts:products ];
+        [[PiwikTracker sharedInstance] ecommerceAddToCart: ecommerceProducts
+                                             currencyCode: options[@"currencyCode"]];
 
         resolve(nil);
     } @catch (NSException *exception) {
@@ -370,7 +389,8 @@ RCT_REMAP_METHOD(trackEcommerceRemoveFromCart,
     @try {
         [self applyOptionalParameters:options];
         EcommerceProducts *ecommerceProducts = [self buildEcommerceProducts:products];
-        [[PiwikTracker sharedInstance] ecommerceRemoveFromCart: ecommerceProducts];
+        [[PiwikTracker sharedInstance] ecommerceRemoveFromCart: ecommerceProducts
+                                                  currencyCode: options[@"currencyCode"]];
 
         resolve(nil);
     } @catch (NSException *exception) {
@@ -400,7 +420,8 @@ RCT_REMAP_METHOD(trackEcommerceOrder,
                                              subTotal:options[@"subTotal"]
                                                   tax:options[@"tax"]
                                              shipping:options[@"shipping"]
-                                             discount:options[@"discount"]];
+                                             discount:options[@"discount"]
+                                         currencyCode:options[@"currencyCode"]];
 
         resolve(nil);
     } @catch (NSException *exception) {
@@ -704,6 +725,24 @@ RCT_REMAP_METHOD(getDispatchInterval,
     }
 }
 
+RCT_REMAP_METHOD(setVisitorIDLifetime,
+                 setVisitorIDLifetimeWithVisitorIDLifetime:(double)visitorIDLifetime
+                 withResolver:(RCTPromiseResolveBlock)resolve
+                 withRejecter:(RCTPromiseRejectBlock)reject)
+{
+    if ([PiwikTracker sharedInstance] == nil) {
+        reject(@"not_initialized", @"Piwik Pro SDK has not been initialized", nil);
+        return;
+    }
+    
+    @try {
+      [PiwikTracker sharedInstance].visitorIDLifetime = visitorIDLifetime;
+        resolve(nil);
+    } @catch (NSException *exception) {
+        reject(exception.name, exception.reason, nil);
+    }
+}
+
 RCT_REMAP_METHOD(setIncludeDefaultCustomVariables,
                  withIncludeDefaultCustomVariables:(BOOL)includeDefaultCustomVariables
                  getDispatchIntervalWithResolver:(RCTPromiseResolveBlock)resolve
@@ -874,6 +913,76 @@ RCT_REMAP_METHOD(isPrefixingOn,
     @try {
         BOOL prefixingEnabled = [PiwikTracker sharedInstance].isPrefixingEnabled;
         resolve(@(prefixingEnabled));
+    } @catch (NSException *exception) {
+        reject(exception.name, exception.reason, nil);
+    }
+}
+
+RCT_REMAP_METHOD(setVisitorIdFromDeepLink,
+                 setVisitorIdFromDeepLinkWithDeepLink:(nonnull NSString*)deepLink
+                 withResolver:(RCTPromiseResolveBlock)resolve
+                 withRejecter:(RCTPromiseRejectBlock)reject)
+{
+    if ([PiwikTracker sharedInstance] == nil) {
+        reject(@"not_initialized", @"Piwik Pro SDK has not been initialized", nil);
+        return;
+    }
+    
+    @try {
+      [[PiwikTracker sharedInstance] setVisitorIdFromDeepLink: deepLink];
+        resolve(nil);
+    } @catch (NSException *exception) {
+        reject(exception.name, exception.reason, nil);
+    }
+}
+
+RCT_REMAP_METHOD(getUserAgent,
+                 getUserAgentWithResolver:(RCTPromiseResolveBlock)resolve
+                 withRejecter:(RCTPromiseRejectBlock)reject)
+{
+    if ([PiwikTracker sharedInstance] == nil) {
+        reject(@"not_initialized", @"Piwik Pro SDK has not been initialized", nil);
+        return;
+    }
+    
+    @try {
+      NSString* userAgent = [PiwikTracker sharedInstance].userAgent;
+        resolve(userAgent);
+    } @catch (NSException *exception) {
+        reject(exception.name, exception.reason, nil);
+    }
+}
+
+RCT_REMAP_METHOD(setSessionHash,
+                 setSessionHashWithSessionHash:(NSInteger)sessionHash
+                 withResolver:(RCTPromiseResolveBlock)resolve
+                 withRejecter:(RCTPromiseRejectBlock)reject)
+{
+    if ([PiwikTracker sharedInstance] == nil) {
+        reject(@"not_initialized", @"Piwik Pro SDK has not been initialized", nil);
+        return;
+    }
+    
+    @try {
+        [[PiwikTracker sharedInstance] setSessionHash:sessionHash];
+        resolve(nil);
+    } @catch (NSException *exception) {
+        reject(exception.name, exception.reason, nil);
+    }
+}
+
+RCT_REMAP_METHOD(getSessionHash,
+                 getSessionHashWithResolver:(RCTPromiseResolveBlock)resolve
+                 withRejecter:(RCTPromiseRejectBlock)reject)
+{
+    if ([PiwikTracker sharedInstance] == nil) {
+        reject(@"not_initialized", @"Piwik Pro SDK has not been initialized", nil);
+        return;
+    }
+    
+    @try {
+        NSInteger sessionHash = [[PiwikTracker sharedInstance] sessionHash];
+        resolve(@(sessionHash));
     } @catch (NSException *exception) {
         reject(exception.name, exception.reason, nil);
     }