+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-## 1.1.0
+## 2.0.0
+
+* Added support for native iOS SDK version 2.1.0
+* Added support for native Android SDK version 2.1.0
+* Fix trackSearch typing
+
+# 1.1.0
 
 * Added support for native iOS SDK version 1.2.1
 * Added support for native Android SDK version 1.2.0